7_forensics/_dev/app/requirements.txt
Flask
gunicorn
Pillow
requests
